+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -267,6 +267,8 @@ public class SampleSQLBuilder {
 						).put(
 							"journalDataFactory", new JournalDataFactory()
 						).put(
+							"layoutDataFactory", new LayoutDataFactory()
+						).put(
 							"messageBoardDataFactory",
 							new MessageBoardDataFactory()
 						).put(
