+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/ddl.ftl
@@ -18,7 +18,7 @@
 		layoutName = "dynamic_data_list_display_" + ddlRecordSetCount
 		portletId = "com_liferay_dynamic_data_lists_web_portlet_DDLDisplayPortlet_INSTANCE_TEST" + ddlRecordSetCount
 
-		layoutModel = dataFactory.newLayoutModel(groupId, layoutName, "", portletId)
+		layoutModel = layoutDataFactory.newLayoutModel(groupId, layoutName, "", portletId)
 	/>
 
 	<@insertLayout _layoutModel=layoutModel />
