+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/fragment.ftl
@@ -6,7 +6,7 @@ ${dataFactory.toInsertSQL(fragmentCollectionModel)}
 
 ${dataFactory.toInsertSQL(fragmentEntryModel)}
 
-<#assign contentLayoutModels = dataFactory.newContentLayoutModels(groupId) />
+<#assign contentLayoutModels = layoutDataFactory.newContentLayoutModels(groupId) />
 
 <#list contentLayoutModels as contentLayoutModel>
 	<@insertContentLayout
